source-dynamodb: mark shard finished only if full shard read (#3418)
When reading streams during backfill, we set a horizon to delay the reading of
change records to avoid older backfill records from being written after a new
change during the horizon window. When doing this, we may read only part of a
shard. If the shard is closed but isn't fully read, we wouldn't return to the
shard later.

This could cause records near the end of the shard to be missed during
backfill.

To fix this, if the shard is only partially read we no longer mark the shard as
finished. Later we will get a new shard iterator after the lastReadSeq id.