source-postgres: Add 'Discover Unpublished Tables'
In https://github.com/estuary/connectors/issues/3335 we removed
the "Discover Only Published Tables" advanced option in favor of
making that behavior unconditionally active.

While that's definitely a more correct default behavior, it turns
out there are (rare) edge cases where we really need the equivalent
of setting that to false. These edge cases are basically ones where
users:

1. Want to manage which tables we're capturing solely via our UI.
2. But cannot use a `FOR ALL TABLES` publication due to data volume.

So this commit is kind of a funhouse-mirror reverse of the previous
one, in that it reinstates the branch around the filtering behavior
but now it's controlled by a new "Discover Unpublished Tables"
advanced option with inverted semantics.

Since we never actually removed the "add unpublished tables to the
publication" logic in the prerequisites, that part should keep on
working as it used to in cases where this new setting is used. We
did also strip out the "autocreate publication if missing" logic as
well in the previous PR, but I think that was still the right call
since it's a one-time setup operation and autocreating it will not
typically be helpful (and not attempting to autocreate it lets us
provide clearer failure messaging for the typical user).