source-dynamics-365-finance-and-operations: fix CSV file path matching
The `metadata.name.startswith` check to determine which CSVs a stream should read was overly broad. For example, the `dimensionattributevalue` stream was trying to read CSVs under `{timestamp_folder}/dimensionattributevalueset`.

Adding a trailing `/` to this check fixes this bug & ensures streams only look for changes in their same named subfolders.

I also corrected incorrect file name formats in the README. The table name is a part of the path, but not part of the actual file name.