source-dynamics-365-finance-and-operations: remove DataType enum
The `DataType` string enum was incomplete and missing a valid data type
(dateTimeOffset). This was causing cryptic errors during discovery like
"Some resource was exhausted.".

I originally included the `DataType` enum when I was planning on using
`SourcedSchema`s to aid when building up the schemas for those messages,
but since the connector ended up not using `SourcedSchema`s, the
`DataType` enum is primarily documentation for `model.json` contents.
This can be sufficiently documented in the `README.md`, and the
connector doesn't need to keep the `DataType` enum around or use it
during validation.