source-postgres-batch: Allow reltuples to be a float
The source column is a `float4` and this appears to be a difference
in behavior between using PGX directly (in source-postgres) and as a
Go sql.DB driver (in source-postgres-batch). In the former case a
non-integer value will be rounded or truncated, but in the latter
it errors out if the value is not an integer, so non-integer reltuples
values are causing errors even though the function is a direct copy-
paste from source-postgres _and_ it works fine in local testing if
the reltuples value is an integer.

At least I'm pretty sure that's what is happening here.