source-braintree-native: extend capture snapshot test duration
Braintree's API takes longer and longer to respond to `effective_date` searches in `disputes` the larger the date window is we request (even though the amount of actual data within the windows is the same). This is causing the capture snapshot to occassionally fail since it doesn't run long enough to capture any `disputes` documents. This commit extends the duration of the capture snapshot test to be long enough for Braintree to respond to our `disputes` search.