source-sqlserver-batch: Fix UUID cursor issue
Splitting capture-value translation from cursor-value translation
meant that UUIDs were getting processed twice (when used a cursors)
and since the byte-shuffling happened in place it was also being
done twice, which was bad. This adds a bit of extra memory alloc-
ation but shouldn't be a big deal, and makes UUID cursors work
correctly again.