materialize-motherduck: add secret scope and check (#3474)
The motherduck materialization creates a temporary secret for operations
to object storage.  It is possible though that an existing permanent
secret will have higher priority than the temporary secret.

With this change, the temporary secret now has a scope set to the
bucket, this increases the priority of the secret for writes to the
bucket.

There could still be a permanent secret that is even more specific,
while unlikely, this could still be true unless we made a temporary
secret for every write/read.  Instead of attempting this, we do a single
check to try to test if our secret will be used and warn if it is
not.

The integration test for Google Cloud was failing with permission
issues, so I regenerated a new HMAC for the service account.