source-postgres: Fix error on missing column in discovery
When a column exists but then is dropped or renamed, it no longer shows
up in discovery. If we still have changes to process with the old name
but it's not present in discovery any more, we error out. But in most
cases we don't actually care about the discovered type name (and in fact
we ought to fix the remaining cases for precisely this reason), so it's
more generally useful to just keep going and hope for the best.