materialize-motherduck: reduce max size for store documents (#3497)
The value of maximum_object_size indirectly affects how much memory
MotherDuck will allocate for parsing JSON.  By bringing this closer to
the actual limit we should be avoid some memory issues.

Our messages are limited to 64MiB, and the document must be less than
this amount to make space for other fields and grpc encoding.  I expect
that the encoded ndjson will be of similar size and also fit within this
limit.