source-outreach: reduce page size after 502 Bad Gateway response
Using the `MAX_PAGE_SIZE` sometimes causes the Outreach API server to
take too long to respond, and the connector ends up receiving a 502 Bad
Gateway response instead (I assume from an intermediate gateway). When
this happens, the connector needs to retry the same request but with a
smaller page size. Eventually, the requested page size will be small
enough that the Outreach API will respond in a reasonable amount of time.