source-klaviyo-native: ignore resources captured in previous sweeps
`fetch_incremental_resources` assumed that using the
`LowerBoundOperator.GREATER_THAN` caused Klaviyo to _not_ return results
whose cursor field was at the lower bound. Turns out, that's not
necessarily true, and the Klaviyo API sometimes returns results that are
on the lower bound when using the greater than operator. We can easily
address this by just ignoring any results received that are not greater
than the log cursor.