sqlcapture+friends: Backfill progress percentages
Refactors the generic SQL CDC main-loop to centralize connectorStatus
logging in one place, and adds logic to compute a backfill progress
percentage and append it to the status text when backfills are in
progress.

The percentage is a "naive" row-count and doesn't take into account
the varying size of different tables. I suspect in general this
will not matter. Also worth noting: the percentage could jump up
or down as tables finish backfilling (because the average progress
of the remaining tables may be different after you remove the one
that just finished), this is why the message is specifically:

    (backfilled 83% of 5 tables)

The hope is that putting the table count that close to the percent
makes it obvious why the progress number changes when the table
count goes down.

To make this work, each database now implements a new method
`EstimatedRowCounts()` which returns our best estimate of the
total size of some tables. The estimate and/or error is cached
so there should be minimal performance impact in any scenario.

If we're unable to get an estimated size for one or more tables
currently backfilling, we currently just give up and report a
simpler count like:

    (3 tables backfilling)

I think this is the right call for a summary statistic, however
it would help if we could break out status by table. There are
some situations where we might have usable estimates for some
but not all tables (Postgres when a table has never been vacuumed
or analyzed).

If the commit message length here didn't clue you in, this is a
surprisingly nuanced bit of code. But I think a single percentage
summary statistic that works most of the time is a good start.