source-braintree-native: increase lookback window to three days for disbursed & disputed transactions
It can take a couple days for transactions' disbursement details to be
available in Braintree, so the connector's strategy of looking back only
a single day results in missing some disbursed transactions.

The connector now looks back three days to avoid missing disbursed
transactions that take a few days to become available. I'm assuming it
can also take a few days for disputed transaction details to be
available, so the lookback window for that search has also been
increased to three days.