filesink: include HTTP client timeout for S3Stores
We've still seen `materialize-s3-iceberg` tasks hang
after merging 02cb9230c6803cb70a46a0bb07bc5079dd7e4e7b. It's still not
clear _why_ the task still hangs, and we need to get a stack
track from an actively hanging materialization to confirm for sure
where it's stuck.

In the meantime in the off chance that the hang is related to not having
a timeout for the HTTP client, I'm setting one to see if it helps stop
`materialize-s3-iceberg` from hanging.