source-hubspot-native: make crm objects use a merge reduction strategy
Change collections for CRM objects to use a merge reduction strategy
instead of the default LWW strategy. The merge reduction strategy is used
at the top level, for the `properties` location, and the `propertiesWithHistory`
location. This ensures that property values and history captured in
calculated property refreshes are merged into the existing values from
previously captured documents. This is part of the "calculated properties
refresh" strategy for keeping calculated properties up to date.