materialize-iceberg: add pagination for list tables (#3513)
AWS glue appears to have a maximum size of 100 tables per response. This adds
pagination over additional pages so that we retrieve the full list of tables. I
have also added pagination for namespaces, though this is harder to test since
we don't have any instances with so many namespaces.