materialize-elasticsearch: allow manual index mapping
Users can now specify the type and format for any mapping, similar to
how it is allowed for the sql materializations.  Only types that we
currently support can be set.  The format parameter can be used with
date fields to allow different formats, such as a long that represents a
unix timestamp.

The method can be used to replicate the current keyword: true option by
setting the type to keyword.  I plan to de-emphasize that method in the
documentation.

It would be nice to allow materializations to have their own
ExistingResource type so that they don't all need to share, in
particular, the ExistingField struct.  This would be a good follow up
change.