materialize-elasticsearch: allow manual index mapping (#3525)
Users can now specify the type and format for any mapping, similar to
how it is allowed for the sql materializations.  Only types that we
currently support can be set.  The format parameter can be used with
date fields to allow different formats, such as a long that represents a
unix timestamp.