source-mongodb: fix race condition checking if streams are caught up
Each time startCatchingUp was called, it would recreate the done
broadcast channel.  If someone was already waiting on this channel then
the channel would never be closed.  Since we call the startCatchingUp
function on a ticker, not syncronously with waiting, this was a race
condition.

With this change the channels are created on demand, and stored so that
the next time all streams are caught up they will be closed.