materialize-sqlserver: use string column size info (#3480)
Create string columns with only enough size for the collection string. If no
size is provided then a MAX size varchar is created. The maxLength in the
schema is validated as the number of codepoints, not bytes, so when creating a
column we reserve 4 bytes for every 1 length if creating a varchar column, or 2
bytes for nvarchar. Note that with nvarchar the size is specified in
byte-pairs. This is almost certainly a massive over count for varchar, since
very few characters would use the full 4 bytes.

A string column is compatible with columns with at least enough space as
required by the schema.

Otherwise column can be migrated to any larger columns size. Migration used the
current strategy of creating a new column, copying the data, and renaming to
the original name.