source-mongodb: fix race condition checking if streams are caught up (#3529)
Each time startCatchingUp was called, the done broadcast channel would be
recreated. If another goroutine was already waiting on this channel then it
would never be closed. Since we call the startCatchingUp function on a ticker,
not synchronously with waiting, this was a race condition.

With this change the channels are created on demand, and stored so that the
next time all streams are caught up they will be closed.