source-hubspot-native: allow custom objects to shadow same named static bindings
Currently, if a custom object and a standard object share the same name,
the connector discovers two resources with that name. This causes
discovery to fail since the runtime errors if two resources are
discovered with the same name.

Now, custom objects will shadow any statically named bindings with the
same name. For example, a `form_submissions` custom object will cause
the connector to _not_ include the standard` form_submissions` resource
in discovery.