source-hubspot-native: introduct useLegacyNamingForCustomObjects flag
If a custom object and standard object have the same name, only a single
resource with that name can be included in discovery. This means the
connector has to exclude one of those objects in the list of discovered
resoures.

To fix this, the `useLegacyNamingForCustomObjects` flag is introduced in
this commit. When selected, the connector uses the current behavior of
using the custom object's name as-is for the resource name, allowing for
custom objects to shadow statically named standard objects. When not
selected, all custom object resources will have their names prepended
with `custom_` to avoid any naming conflicts between custom and standard
objects.

The default for `useLegacyNamingForCustomObjects` is `true` to maintain
the current behavior for all existing captures. Once I update all
existing captures to have `useLegacyNamingForCustomObjects` explicitly
set to `true`, I'll change the connector's default to `false` so all
new captures will prepend custom object resource names with `custom_`.