materialize-snowflake: interleave document loading (#3556)
In the current code, queries are started concurrently but the results
are processed sequentially.  For large result sets, this could result in
a long period before the results start to be read.  My theory is that
this can cause the connection to time out since we leave the response
idle for too long.

With this change, the result of all active queries are read concurrently
using a channel and a single loader goroutine instead of the mutex to
serialize message output.