s/jsonschema_extra/jsonschema_extras/g
As far as I can tell, `jsonschema_extra` is not a valid annotation
and is ignored by the JSON schema reflection package we use. This
doesn't seem to make a difference in practice (connectors with it
set correctly vs missing seem to render identically in the UI) but
we probably ought to fix it just to keep things tidy.