materialize-snowflake: retry load queries
If an error occurs while running a load query, we would return an error
and start over on the next run.  This can be a problem as you scale up
the size of the transaction, as it becomes more likely for an error to
occur and more expensive to rerun the transaction from the start.

In order to retry a query after it fails, we need to resume at the same
position without loading documents multiple times.  I added an ORDER BY
clause to ensure the results have the same ordering, and a row number so
we can skip over results we already processed.