materialize-s3-iceberg: use 1 based indexing for field-id (#3606)
Iceberg generally uses 1-based indexing for field-ids, and adjusts this value
forward automatically when writing the schema.  However, this adjustment does
not apply to the schema.name_mapping which is now attached when creating the
initial table.  This could cause a discrepancy between the field-id making
preventing the table from being loaded.