source-http-ingest: update for breaking changes in flow deps
The flow `json` and `doc` crates had some breaking changes. Most were pretty
simple. The main issue was that a `serde_json::de::Deserializer` can no longer
be used to validate a serialized json document.  The fix for that is to fully
parse documents as `serde_json::Value`s, and keeps them in that form until
after validation is completed.