source-iterable: stream and incrementally process `events` API responses
We've observed `source-iterable` captures OOMing when processing the
`events` stream. I suspect this is happening when a single API response
is too large to fit in memory. Switching the `events` stream to
stream and incrementally process the API response instead of loading it
all in memory should help avoid those OOMs.