sqlcapture: Raise diagnostics timeout slightly
Increases the threshold for the "automated replication diagnostics"
triggering from one minute to five minutes. Generally if it takes
less than 5 minutes there's still no cause for concern, we probably
just had a bit of a replication backlog to work through. And in fact
running diagnostics on too much of a hair-trigger just makes it
harder to see what's actually going on with a capture.