source-sqlserver: Lowercase sys.partitions in backfill stats
This is kind of the inverse of the usual `information_schema` vs
`INFORMATION_SCHEMA` capitalization issue we see with discovery
queries. The canonical names of the `sys.foo` tables are lower-
case and if we capitalize them then things will work fine in the
default collation but in some locales with tricky case folding
rules (like Turkish 'i') `'SYS.PARTITIONS' != 'sys.partitions'`
and the query will fail.

Using lowercase is always more correct.