source-iterable: retry connection timeout errors during `events` stream
The `events` stream occassionally errors because the Iterable API doesn't
respond at all. We've seen this behavior before with other APIs and handle it
alright with the CDK's `http` module, but imported connectors don't use
that module. So, to reduce the timeout from a few hours to something more
reasonable, we need to include specific timeout values in the
`request_kwargs` method. And to retry after a timeout, we can wrap the
`read_records` calls with some retry logic.

If reading records for a specific user email fails more than `max_retries`
times, we skip syncing that record on this sweep. That's ok because
`events` is a full refresh stream and will try to sync that record on
the next sweep anyway.