source-chargebee-native: Enforce proper filter use and sorting behavior for incremental resources
Enforces the use of "[between]" filter and removes "[before]" and "[after]" since the between filter is inclusive the the only one used. The after and before filters are exclusive and more difficult to use in our context.

Ensures the use of "sort_by" in the API request that support it.