source-iterable: handle edge case in AdjustableSliceGenerator
When the amount of time spent on a slice is extremely short, the `days_per_minute` variable could be infinity & the connector would fail with an `OverflowError: cannot convert float infinity to integer` exception when trying to calculate the `next_range`.

When this edge case happens, it's safe to just use the DEFAULT_RANGE_DAYS for the next slice's range.