source-s3: add aws iam support (#3645)
Using IAM required a newer flow-parser binary, since the older one had a schema
issue that was fixed in estuary/flow#2567.

The packaging in flow has been updated, the release name is different and the
files are packaged individually.  I updated fetch-flow.sh to get some of the
new binaries, but the old package had some files no longer available, so we
probably need to think about what the best way to get these binaries is.