source-iterable: remove `events` stream
No active captures have successfully completed a sync of the `events`
stream, even when it's been left to run for over half a day. The strategy
of making a request for each user isn't scalable - it will take too long
to complete an entire sync and too much data (50+ GB) can be stored on
disk before checkpointing any of it.

Since this stream hasn't ever worked for us and implementing a
different, truly incremental strategy will require re-writing as a
native connector, I'd like to remove `events` from the imported
`source-iterable` connector to avoid the undesirable failure mode of
running reactors out of disk space.